refactor : 패키지 이름 변경 및 유효성 검증 메서드 추가
