+++ src/main/java/roomescape/entity/Reservation.java
@@ -1,8 +1,9 @@
-package roomescape.dto;
+package roomescape.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
+import roomescape.exception.InvalidException;
 
 public class Reservation {
 
@@ -14,8 +15,11 @@ public class Reservation {
 
     public Reservation(long id, String name, LocalDate date, LocalTime time) {
         this.id = id;
+        validateName(name);
         this.name = name;
+        validateDate(date);
         this.date = date;
+        validateTime(time);
         this.time = time;
     }
 
@@ -33,6 +37,24 @@ public class Reservation {
 
     public LocalTime getTime() {
         return time;
+    }
+
+    private void validateName(String name) {
+        if (name == null || name.isBlank()) {
+            throw new InvalidException("Name is required");
+        }
+    }
+
+    private void validateDate(LocalDate date) {
+        if (date == null) {
+            throw new InvalidException("Date is required");
+        }
+    }
+
+    private void validateTime(LocalTime time) {
+        if (time == null) {
+            throw new InvalidException("Time is required");
+        }
     }
 }
